backend/llm_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
openai==1.12.0
httpx==0.27.0
groq==0.4.2
anthropic==0.18.1
PyPDF2==3.0.1
python-multipart==0.0.9 